edit

+++ userportal/src/main/java/com/bookstore/userportal/domain/CouponImport.java
@@ -23,13 +23,13 @@ public class CouponImport {
     private LocalDate dateImport;
 
     @Column(name = "price_book")
-    private Double priceBook;
+    private Integer priceBook;
 
     @Column(name = "number_import")
     private Integer numberImport;
 
     @Column(name = "total_price")
-    private Double totalPrice;
+    private Integer totalPrice;
 
     @ManyToOne
     @JoinColumn(name = "book_id")
